[INTERNAL] Split code of user model into several files
