+++ collectives/utils/payment.py
@@ -55,7 +55,9 @@ def extract_payments(event_id=None, page=None, pagesize=50, filters=None):
                         f"payment_list: {value} cannot be converted to an int"
                     )
             elif field == "item.event.event_type.name" and value is not None:
+                # pylint: disable=comparison-with-callable
                 query = query.filter(Event.event_type_id == int(value))
+                # pylint: enable=comparison-with-callable
             elif field == "item.title":
                 query = query.filter(PaymentItem.title.like(f"%{value}%"))
             elif field == "price.title":
